pyproject.toml
[tool.poetry]
name = "cfb8"
version = "0.1.0"
description = ""
authors = ["knqyf263 <knqyf263@gmail.com>"]
license = "mit"

[tool.poetry.dependencies]
python = "^3.8"
pycryptodome = "^3.9.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
